[BE] refactor: LogProperty 패키지를 property로 변경 및 클래스 이동 #650
- `LogProperty` 관련 클래스들을 새로운 패키지 구조로 이동
- Hibernate 관련 설정 값 변경 (e.g., `HibernateQueryInspector`)
