+++ backend/app/src/main/java/com/onair/hearit/app/auth/application/AuthService.java
@@ -1,24 +1,24 @@
 package com.onair.hearit.app.auth.application;
 
-import com.onair.hearit.app.auth.domain.RefreshToken;
 import com.onair.hearit.app.auth.dto.request.LoginRequest;
 import com.onair.hearit.app.auth.dto.request.OAuthLoginRequest;
 import com.onair.hearit.app.auth.dto.request.SignupRequest;
 import com.onair.hearit.app.auth.dto.response.LoginTokenResponse;
 import com.onair.hearit.app.auth.dto.response.OAuthUserInfoResponse;
 import com.onair.hearit.app.auth.infrastructure.jwt.JwtTokenProvider;
-import com.onair.hearit.app.auth.infrastructure.repository.RefreshTokenRepository;
 import com.onair.hearit.app.exception.custom.InvalidInputException;
 import com.onair.hearit.app.exception.custom.NotFoundException;
 import com.onair.hearit.app.exception.custom.UnauthorizedException;
 import com.onair.hearit.core.domain.Member;
 import com.onair.hearit.core.domain.OAuthProvider;
+import com.onair.hearit.core.domain.RefreshToken;
 import com.onair.hearit.core.infrastructure.jpa.MemberRepository;
-import com.onair.hearit.core.log.dto.logproperty.auth.LoginLogProperty;
-import com.onair.hearit.core.log.dto.logproperty.auth.SignUpLogProperty;
-import com.onair.hearit.core.log.dto.logproperty.auth.TokenRefreshLogProperty;
-import com.onair.hearit.core.log.dto.logproperty.auth.WithdrawalLogProperty;
+import com.onair.hearit.core.infrastructure.jpa.RefreshTokenRepository;
 import com.onair.hearit.core.log.logger.JsonLogger;
+import com.onair.hearit.core.log.property.auth.LoginLogProperty;
+import com.onair.hearit.core.log.property.auth.SignUpLogProperty;
+import com.onair.hearit.core.log.property.auth.TokenRefreshLogProperty;
+import com.onair.hearit.core.log.property.auth.WithdrawalLogProperty;
 import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
