+++ backend/app/src/main/java/com/onair/hearit/app/auth/infrastructure/jwt/JwtAuthenticationFilter.java
@@ -3,9 +3,9 @@ package com.onair.hearit.app.auth.infrastructure.jwt;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.onair.hearit.app.auth.domain.RequestUser;
 import com.onair.hearit.app.exception.ErrorCode;
-import com.onair.hearit.core.log.dto.logproperty.api.ExceptionLogProperty;
 import com.onair.hearit.core.log.logger.ConsoleLogger;
 import com.onair.hearit.core.log.logger.JsonLogger;
+import com.onair.hearit.core.log.property.api.ExceptionLogProperty;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
