+++ backend/app/src/test/java/com/onair/hearit/app/auth/presentation/AuthIntegrationTest.java
@@ -3,19 +3,19 @@ package com.onair.hearit.app.auth.presentation;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import com.onair.hearit.app.auth.domain.RefreshToken;
 import com.onair.hearit.app.auth.dto.request.LoginRequest;
 import com.onair.hearit.app.auth.dto.request.SignupRequest;
 import com.onair.hearit.app.auth.dto.request.TokenReissueRequest;
 import com.onair.hearit.app.auth.dto.response.LoginTokenResponse;
 import com.onair.hearit.app.auth.dto.response.TokenReissueResponse;
 import com.onair.hearit.app.auth.infrastructure.jwt.JwtTokenProvider;
-import com.onair.hearit.app.auth.infrastructure.repository.RefreshTokenRepository;
 import com.onair.hearit.app.fixture.DbHelper;
 import com.onair.hearit.app.fixture.IntegrationTest;
 import com.onair.hearit.core.domain.Member;
+import com.onair.hearit.core.domain.RefreshToken;
 import com.onair.hearit.core.fixture.TestFixture;
 import com.onair.hearit.core.infrastructure.jpa.MemberRepository;
+import com.onair.hearit.core.infrastructure.jpa.RefreshTokenRepository;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import java.time.LocalDateTime;
