+++ backend/boot/src/main/resources/application-dev.properties
@@ -33,7 +33,7 @@ spring.jackson.time-zone=Asia/Seoul
 # logging
 logging.config=classpath:log4j2/log4j2.xml
 server.forward-headers-strategy=native
-spring.jpa.properties.hibernate.session_factory.statement_inspector=com.onair.hearit.core.log.HibernateQueryInspector
+spring.jpa.properties.hibernate.session_factory.statement_inspector=com.onair.hearit.core.log.dbtrace.HibernateQueryInspector
 # Flyaway
 spring.flyway.user=${MYSQL_MIGRATION_USER}
 spring.flyway.password=${MYSQL_MIGRATION_PASSWORD}
