+++ backend/core/src/main/java/com/onair/hearit/core/log/ApiLoggingAspect.java
@@ -1,10 +1,10 @@
 package com.onair.hearit.core.log;
 
-import com.onair.hearit.core.log.dto.logproperty.api.ExceptionLogProperty;
-import com.onair.hearit.core.log.dto.logproperty.api.RequestLogProperty;
-import com.onair.hearit.core.log.dto.logproperty.api.ResponseLogProperty;
 import com.onair.hearit.core.log.logger.ConsoleLogger;
 import com.onair.hearit.core.log.logger.JsonLogger;
+import com.onair.hearit.core.log.property.api.ExceptionLogProperty;
+import com.onair.hearit.core.log.property.api.RequestLogProperty;
+import com.onair.hearit.core.log.property.api.ResponseLogProperty;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.Arrays;
 import java.util.Optional;
