+++ backend/core/src/main/java/com/onair/hearit/core/log/logger/ConsoleLogger.java
@@ -1,9 +1,9 @@
 package com.onair.hearit.core.log.logger;
 
-import com.onair.hearit.core.log.dto.logproperty.api.ExceptionLogProperty;
-import com.onair.hearit.core.log.dto.logproperty.api.RequestLogProperty;
-import com.onair.hearit.core.log.dto.logproperty.api.ResponseLogProperty;
 import com.onair.hearit.core.log.formatter.ConsoleLogFormatter;
+import com.onair.hearit.core.log.property.api.ExceptionLogProperty;
+import com.onair.hearit.core.log.property.api.RequestLogProperty;
+import com.onair.hearit.core.log.property.api.ResponseLogProperty;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.stereotype.Component;
