+++ backend/core/src/main/java/com/onair/hearit/core/log/property/auth/TokenRefreshLogProperty.java
@@ -1,7 +1,7 @@
-package com.onair.hearit.core.log.dto.logproperty.auth;
+package com.onair.hearit.core.log.property.auth;
 
-import com.onair.hearit.core.log.dto.LogEvent;
-import com.onair.hearit.core.log.dto.logproperty.LogProperty;
+import com.onair.hearit.core.log.LogEvent;
+import com.onair.hearit.core.log.property.LogProperty;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import lombok.AccessLevel;
